ask for filename if none is given on command line
I had to rearrange some functions and also the place where the tk root window
is created. Otherwise the dialog that asks for a filename will not work.
Also the global dictionaries now have a function that sets them. This script
is getting really long. It should be made into an object oriented app.

I also added menu items to specify the columns we use in each file.