Assign slightly different times to consecutive data blocks with same timelabel
When we read text file data (in function append_textfile_data) and find
several consecutive data blocks with the same timelabel, we now assign
slightly different times to each of these data blocks to make it easer for
the viewer to graph these data blocks. This small artificial change in
time is produced by multiplying the actual time by 1.00000000000001.